# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFtaXJleDEyOEBnbWFpbC5jb20iLCJleHAiOjUyNjgzMTg5OTYsImV4cGlyZV9hdCI6IiIsImZpcnN0bmFtZSI6Itin2YXbjNixIiwiaWQiOjEsImxhc3RuYW1lIjoi2LTbjNix2K_ZhNuMIiwibW9iaWxlIjoiMDkwMjQ4MDk3NTAiLCJvcmlnX2lhdCI6MTY2ODMyMjU5Niwic3RhdHVzIjoiIn0.x7BKuxw288cm1JsskGRD178UPmNz-xRwkWHtb0WsU74",
        "title": "Selloora Backend API",
        "contact": {
            "name": "API Support",
            "url": "https://www.amirshirdel.ir",
            "email": "amirex128@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8585",
    "basePath": "/",
    "paths": {
...
        "/user/product": {
            "get": {
                "description": "محصولات میتوانند خصوصیات مختلفی داشته باشند که این خصوصیت ها شامل رنگ و اندازه میباشد",
                "tags": [
                    "product"
                ],
                "summary": "لیست محصولات",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "متن جستجو",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "شماره صفحه",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "تعداد صفحه",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "مرتب سازی براساس desc/asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/user/product/create": {
            "post": {
                "description": "محصولات میتوانند خصوصیات مختلفی داشته باشند که این خصوصیت ها شامل رنگ و اندازه میباشد",
                "tags": [
                    "product"
                ],
                "summary": "ایجاد محصول",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.CreateProduct"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/product/delete/{id}": {
            "post": {
                "description": "محصولات میتوانند خصوصیات مختلفی داشته باشند که این خصوصیت ها شامل رنگ و اندازه میباشد",
                "tags": [
                    "product"
                ],
                "summary": "حذف محصول",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه محصول",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/product/show/{id}": {
            "get": {
                "description": "محصولات میتوانند خصوصیات مختلفی داشته باشند که این خصوصیت ها شامل رنگ و اندازه میباشد",
                "tags": [
                    "product"
                ],
                "summary": "نمایش محصول",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "شناسه محصول",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/product/update": {
            "post": {
                "description": "محصولات میتوانند خصوصیات مختلفی داشته باشند که این خصوصیت ها شامل رنگ و اندازه میباشد",
                "tags": [
                    "product"
                ],
                "summary": "ویرایش محصول",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authentication",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ورودی",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/DTOs.UpdateProduct"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "DTOs.CreateProduct": {
            "type": "object",
            "required": [
                "categoryID",
                "gallery_ids",
                "name",
                "price",
                "quantity",
                "shopID"
            ],
            "properties": {
                "categoryID": {
                    "description": "شناسه دسته بندی",
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "description": "توضیحات محصول",
                    "type": "string",
                    "example": "توضیحات محصول"
                },
                "endedAt": {
                    "description": "تاریخ پایان فروش",
                    "type": "string",
                    "example": "2025-01-01 00:00:00"
                },
                "gallery_ids": {
                    "description": "شناسه تصاویر انتخاب شده برای این محصول",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "manufacturer": {
                    "description": "برند محصول",
                    "type": "string",
                    "example": "سامسونگ"
                },
                "name": {
                    "description": "نام محصول",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "گوشی موبایل گلگسی نوت ۱۰"
                },
                "optionId": {
                    "description": "شناسه آشپن محصول",
                    "type": "integer",
                    "example": 1
                },
                "optionItemID": {
                    "description": "شناسه آیتم انتخاب شده از آپشن",
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "description": "قیمت محصول",
                    "type": "number",
                    "example": 1000000
                },
                "quantity": {
                    "description": "تعداد موجودی",
                    "type": "integer",
                    "example": 10
                },
                "shopID": {
                    "description": "شناسه فروشگاه",
                    "type": "integer",
                    "example": 0
                },
                "startedAt": {
                    "description": "تاریخ شروع فروش",
                    "type": "string",
                    "example": "2020-01-01 00:00:00"
                }
            }
        },
...
        "DTOs.UpdateProduct": {
            "type": "object",
            "required": [
                "categoryID"
            ],
            "properties": {
                "active": {
                    "description": "وضعیت فعال یا غیر فعال بودن محصول",
                    "type": "boolean",
                    "example": true
                },
                "categoryID": {
                    "description": "شناسه دسته بندی",
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "description": "توضیحات محصول",
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 3,
                    "example": "توضیحات محصول"
                },
                "endedAt": {
                    "description": "تاریخ پایان فروش",
                    "type": "string",
                    "example": "2025-01-01 00:00:00"
                },
                "gallery_ids": {
                    "description": "شناسه تصاویر انتخاب شده برای این محصول",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "description": "شناسه محصول برای ویرایش",
                    "type": "integer",
                    "example": 1
                },
                "manufacturer": {
                    "description": "برند محصول",
                    "type": "string",
                    "example": "سامسونگ"
                },
                "name": {
                    "description": "نام محصول",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "گوشی موبایل گلگسی نوت ۱۰"
                },
                "optionId": {
                    "description": "شناسه آیتم انتخاب شده از آپشن",
                    "type": "integer",
                    "example": 1
                },
                "optionItemID": {
                    "description": "شناسه آیتم انتخاب شده از آپشن",
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "description": "قیمت محصول",
                    "type": "number",
                    "example": 1000000
                },
                "quantity": {
                    "description": "تعداد موجودی",
                    "type": "integer",
                    "example": 10
                },
                "shop_id": {
                    "description": "شناسه فروشگاه جهت انتقال یک محصول از یک فروشگاه به فروشگاه دیگر",
                    "type": "integer",
                    "example": 2
                },
                "startedAt": {
                    "description": "تاریخ شروع فروش",
                    "type": "string",
                    "example": "2020-01-01 00:00:00"
                }
            }
        },
...
    }
}